requirement.txt
google-generativeai
faiss-cpu
sentence-transformers
numpy
python-dotenv
streamlit
replicate

